+++ packages/lumx-angularjs/src/components/dialog/dialog_directive.js
@@ -52,15 +52,6 @@ function DialogController(
     const _dialog = $element;
 
     /**
-     * The dialog overlay.
-     *
-     * @type {Element}
-     */
-    const _dialogOverlay = angular.element('<div/>', {
-        class: `${CSS_PREFIX}-dialog-overlay`,
-    });
-
-    /**
      * The event scheduler id.
      *
      * @type {string}
@@ -80,6 +71,13 @@ function DialogController(
      * @type {boolean}
      */
     let _isConfirmDialog = false;
+
+    /**
+     * Whether the sentinels are intersecting with dialog content or not.
+     *
+     * @type {Object}
+     */
+    const _isIntersecting = { top: false, bottom: false };
 
     /**
      * The dialog parent element.
@@ -139,7 +137,6 @@ function DialogController(
 
         $rootScope.$broadcast('lx-dialog__close-start', lx.id, canceled, params);
 
-        _dialogOverlay.addClass(`${CSS_PREFIX}-dialog-overlay--is-hidden`);
         _dialog.addClass(`${CSS_PREFIX}-dialog--is-hidden`);
 
         if (angular.isDefined(_sourceEl)) {
@@ -147,11 +144,6 @@ function DialogController(
         }
 
         $timeout(() => {
-            _dialogOverlay
-                .off('click', _close)
-                .removeClass(`${CSS_PREFIX}-dialog-overlay--is-hidden`)
-                .remove();
-
             if (_isAlertDialog || _isConfirmDialog) {
                 _dialog.remove();
             } else {
@@ -169,6 +161,15 @@ function DialogController(
     }
 
     /**
+     * Remove dialog scroll state if the sentinels are not intersecting.
+     */
+    function _removeScrollState() {
+        if (!_isIntersecting.top && !_isIntersecting.bottom) {
+            _dialog.removeClass(`${CSS_PREFIX}-dialog--has-scroll`);
+        }
+    }
+
+    /**
      * Create dialog content observer.
      */
     function _createObserver() {
@@ -178,23 +179,45 @@ function DialogController(
 
         const observer = new IntersectionObserver((entries) => {
             entries.forEach((entry) => {
-                if (entry.target.classList.contains(`${CSS_PREFIX}-dialog__sentinel--top`) && !lx.forceHeaderDivider) {
-                    if (entry.isIntersecting) {
-                        dialogHeader.removeClass(`${CSS_PREFIX}-dialog__header--has-divider`);
-                    } else {
-                        dialogHeader.addClass(`${CSS_PREFIX}-dialog__header--has-divider`);
+                if (entry.target.classList.contains(`${CSS_PREFIX}-dialog__sentinel--wrapper`)) {
+                    if (!entry.isIntersecting) {
+                        _dialog.addClass(`${CSS_PREFIX}-dialog--has-scroll`);
                     }
                 }
 
-                if (
-                    entry.target.classList.contains(`${CSS_PREFIX}-dialog__sentinel--bottom`) &&
-                    !lx.forceFooterDivider
-                ) {
+                if (entry.target.classList.contains(`${CSS_PREFIX}-dialog__sentinel--top`)) {
                     if (entry.isIntersecting) {
-                        dialogFooter.removeClass(`${CSS_PREFIX}-dialog__footer--has-divider`);
+                        _isIntersecting.top = false;
+                        _removeScrollState();
+
+                        if (!lx.forceHeaderDivider) {
+                            dialogHeader.removeClass(`${CSS_PREFIX}-dialog__header--has-divider`);
+                        }
+                    } else {
+                        _isIntersecting.top = true;
+
+                        if (!lx.forceHeaderDivider) {
+                            dialogHeader.addClass(`${CSS_PREFIX}-dialog__header--has-divider`);
+                        }
+                    }
+                }
+
+                if (entry.target.classList.contains(`${CSS_PREFIX}-dialog__sentinel--bottom`)) {
+                    if (entry.isIntersecting) {
+                        _isIntersecting.bottom = false;
+                        _removeScrollState();
+
+                        if (!lx.forceFooterDivider) {
+                            dialogFooter.removeClass(`${CSS_PREFIX}-dialog__footer--has-divider`);
+                        }
+
                         $rootScope.$broadcast('lx-dialog__scroll-end', lx.id);
                     } else {
-                        dialogFooter.addClass(`${CSS_PREFIX}-dialog__footer--has-divider`);
+                        _isIntersecting.bottom = true;
+
+                        if (!lx.forceFooterDivider) {
+                            dialogFooter.addClass(`${CSS_PREFIX}-dialog__footer--has-divider`);
+                        }
                     }
                 }
             });
@@ -228,23 +251,14 @@ function DialogController(
             return;
         }
 
-        LxDepthService.increase();
-
-        _dialogOverlay
-            .css('z-index', LxDepthService.get())
-            .appendTo('body')
-            .show();
-
-        if (angular.isUndefined(lx.autoClose) || lx.autoClose) {
-            _dialogOverlay.on('click', _close);
-        }
-
         if (angular.isUndefined(lx.escapeClose) || lx.escapeClose) {
             _idEventScheduler = LxEventSchedulerService.register('keydown', _onKeyDown);
         }
 
+        LxDepthService.increase();
+
         _dialog
-            .css('z-index', LxDepthService.get() + 1)
+            .css('z-index', LxDepthService.get())
             .appendTo('body')
             .addClass(`${CSS_PREFIX}-dialog--is-shown`);
 
@@ -296,9 +310,19 @@ function DialogController(
         return classes;
     }
 
+    /**
+     * Close dialog on overlay click.
+     */
+    function onOverlayClick() {
+        if (angular.isUndefined(lx.autoClose) || lx.autoClose) {
+            _close();
+        }
+    }
+
     /////////////////////////////
 
     lx.getClasses = getClasses;
+    lx.onOverlayClick = onOverlayClick;
 
     /////////////////////////////
     //                         //